# Use objectpath.Encoder to amortize the cost of encoding the paths of multiple objects